+++ go.mod
@@ -5,8 +5,8 @@ go 1.21
 require (
 	github.com/senzing/g2-sdk-go v0.6.8
 	github.com/senzing/go-cmdhelping v0.1.8
-	github.com/senzing/go-common v0.2.14
-	github.com/senzing/go-databasing v0.2.8
+	github.com/senzing/go-common v0.3.0
+	github.com/senzing/go-databasing v0.2.9
 	github.com/senzing/go-logging v1.3.2
 	github.com/senzing/go-observing v0.2.7
 	github.com/senzing/go-sdk-abstract-factory v0.3.2
@@ -28,7 +28,7 @@ require (
 	github.com/lib/pq v1.10.9 // indirect
 	github.com/magiconair/properties v1.8.7 // indirect
 	github.com/mattn/go-sqlite3 v1.14.17 // indirect
-	github.com/microsoft/go-mssqldb v1.5.0 // indirect
+	github.com/microsoft/go-mssqldb v1.6.0 // indirect
 	github.com/mitchellh/mapstructure v1.5.0 // indirect
 	github.com/pelletier/go-toml/v2 v2.1.0 // indirect
 	github.com/russross/blackfriday/v2 v2.1.0 // indirect
