+++ senzingschema/senzingschema_test.go
@@ -108,105 +108,3 @@ func TestSenzingSchemaImpl_UnregisterObserver(test *testing.T) {
 	testObject.InitializeSenzing(ctx)
 	testObject.UnregisterObserver(ctx, observer1)
 }
-
-// ----------------------------------------------------------------------------
-// Examples for godoc documentation
-// ----------------------------------------------------------------------------
-
-func ExampleSenzingSchemaImpl_InitializeSenzing() {
-	// For more information, visit https://github.com/Senzing/init-database/blob/main/senzingschema/senzingschema_test.go
-	ctx := context.TODO()
-	senzingEngineConfigurationJson, err := g2engineconfigurationjson.BuildSimpleSystemConfigurationJsonUsingEnvVars()
-	if err != nil {
-		fmt.Print(err)
-	}
-	senzingSchema := &SenzingSchemaImpl{
-		SenzingEngineConfigurationJson: senzingEngineConfigurationJson,
-	}
-	err = senzingSchema.SetLogLevel(ctx, logging.LevelInfoName)
-	if err != nil {
-		fmt.Println(err)
-	}
-	err = senzingSchema.InitializeSenzing(ctx)
-	if err != nil {
-		fmt.Println(err)
-	}
-	// Output:
-}
-
-func ExampleSenzingSchemaImpl_RegisterObserver() {
-	// For more information, visit https://github.com/Senzing/init-database/blob/main/senzingschema/senzingschema_test.go
-	ctx := context.TODO()
-	anObserver := &observer.ObserverNull{
-		Id:       "Observer 1",
-		IsSilent: true,
-	}
-	senzingEngineConfigurationJson, err := g2engineconfigurationjson.BuildSimpleSystemConfigurationJsonUsingEnvVars()
-	if err != nil {
-		fmt.Print(err)
-	}
-	senzingSchema := &SenzingSchemaImpl{
-		SenzingEngineConfigurationJson: senzingEngineConfigurationJson,
-	}
-	err = senzingSchema.RegisterObserver(ctx, anObserver)
-	if err != nil {
-		fmt.Print(err)
-	}
-	// Output:
-}
-
-func ExampleSenzingSchemaImpl_SetLogLevel() {
-	// For more information, visit https://github.com/Senzing/init-database/blob/main/senzingschema/senzingschema_test.go
-	ctx := context.TODO()
-	senzingEngineConfigurationJson, err := g2engineconfigurationjson.BuildSimpleSystemConfigurationJsonUsingEnvVars()
-	if err != nil {
-		fmt.Print(err)
-	}
-	senzingSchema := &SenzingSchemaImpl{
-		SenzingEngineConfigurationJson: senzingEngineConfigurationJson,
-	}
-	err = senzingSchema.SetLogLevel(ctx, logging.LevelInfoName)
-	if err != nil {
-		fmt.Print(err)
-	}
-	// Output:
-}
-
-func ExampleSenzingSchemaImpl_SetObserverOrigin() {
-	// For more information, visit https://github.com/Senzing/init-database/blob/main/senzingschema/senzingschema_test.go
-	ctx := context.TODO()
-	senzingEngineConfigurationJson, err := g2engineconfigurationjson.BuildSimpleSystemConfigurationJsonUsingEnvVars()
-	if err != nil {
-		fmt.Print(err)
-	}
-	senzingSchema := &SenzingSchemaImpl{
-		SenzingEngineConfigurationJson: senzingEngineConfigurationJson,
-	}
-	senzingSchema.SetObserverOrigin(ctx, "TestObserver")
-	// Output:
-}
-
-func ExampleSenzingSchemaImpl_UnregisterObserver() {
-	// For more information, visit https://github.com/Senzing/init-database/blob/main/senzingschema/senzingschema_test.go
-	ctx := context.TODO()
-	anObserver := &observer.ObserverNull{
-		Id:       "Observer 1",
-		IsSilent: true,
-	}
-	senzingEngineConfigurationJson, err := g2engineconfigurationjson.BuildSimpleSystemConfigurationJsonUsingEnvVars()
-	if err != nil {
-		fmt.Print(err)
-	}
-	senzingSchema := &SenzingSchemaImpl{
-		SenzingEngineConfigurationJson: senzingEngineConfigurationJson,
-	}
-	err = senzingSchema.RegisterObserver(ctx, anObserver)
-	if err != nil {
-		fmt.Print(err)
-	}
-	err = senzingSchema.UnregisterObserver(ctx, anObserver)
-	if err != nil {
-		fmt.Print(err)
-	}
-	// Output:
-}
